Fix compatibility issues for sbt plugins
